Inimigos andando, tiro funcionando porém saindo errado o ângulo

+++ Project3D/headers/utils.h
@@ -10,17 +10,17 @@
 #include <math.h>
 
 struct vec3 {
-	GLfloat x, y, z;
+	GLdouble x, y, z;
 	vec3() :x(0.0), y(0.0), z(0.0) {}
-	vec3(GLfloat x_, GLfloat y_, GLfloat z_) :x(x_), y(y_), z(z_) {}
+	vec3(GLdouble x_, GLdouble y_, GLdouble z_) :x(x_), y(y_), z(z_) {}
 	vec3 operator+(const vec3& v)const { return vec3(x + v.x, y + v.y, z + v.z); }
-	vec3 operator*(GLfloat k) { return vec3(x*k, y*k, z*k); }
+	vec3 operator*(GLdouble k) { return vec3(x*k, y*k, z*k); }
 };
 
 struct vec2 {
-	GLfloat x, y;
+	GLdouble x, y;
 	vec2(): x(0.0), y(0.0) {}
-	vec2(GLfloat x_, GLfloat y_) :x(x_), y(y_) {}
+	vec2(GLdouble x_, GLdouble y_) :x(x_), y(y_) {}
 };
